+++ app/Http/Controllers/Backend/SubCategoryController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers\Backend;
 use App\DataTables\SubCategoryDataTable;
 use App\Http\Controllers\Controller;
 use App\Models\Category;
+use App\Models\ChildCategory;
 use App\Models\SubCategory;
 use Illuminate\Http\Request;
 use PhpParser\Node\Stmt\Return_;
@@ -96,6 +97,10 @@ class SubCategoryController extends Controller
     public function destroy(string $id)
     {
         $subCategory = SubCategory::findOrFail($id);
+        $childCategory = ChildCategory::where('sub_category_id', $subCategory->id)->count();
+        if ($childCategory > 0) {
+            return response(['status' => 'error', 'message' => 'This Item Contain sub Items, you have to delete the sub items first']);
+        }
         $subCategory->delete();
         return response(['status' => 'success', 'message' => 'Item Deleted Succefully']);
     }
